+++ CHANGES.md
@@ -126,6 +126,7 @@
     -   Fix bug with adjacent bold font-locking in a list
         item. ([GH-176][])
     -   Prevent matching italics, bold, and inline code in comments.
+    -   Prevent matching italics and bold in URLs.
 
   [gh-123]: https://github.com/jrblevin/markdown-mode/issues/123
   [gh-130]: https://github.com/jrblevin/markdown-mode/issues/130